vertical/horizontal picker split in 2 components
