+++ src/app/components/linear-picker/linear-picker-common.scss
@@ -1,0 +1,92 @@
+.linear-picker {
+    font-family: 'Roboto', sans-serif;
+    position: fixed;
+    z-index: 9999;
+    -webkit-user-select: none;
+    user-select: none;
+    padding: 0;
+    background: #000;
+    color: #fff;
+    border: 2px solid #fff;
+    width: auto;
+    min-width: max-content;
+    white-space: nowrap;
+}
+
+.value-cell {
+    position: relative;
+    font-size: 0.8em;
+    text-align: center;
+    cursor: pointer;
+    transition: background-color 0.15s ease-in-out;
+    flex-shrink: 0;
+}
+
+.value-cell.disabled {
+    cursor: not-allowed;
+    color: #666;
+    opacity: 0.6;
+}
+
+.value-cell.selected {
+    background: var(--bt-yellow-background-bright);
+    font-weight: bold;
+    z-index: 2;
+}
+
+.state-button-cell.highlight:not(.disabled),
+.button-cell.highlight:not(.disabled) {
+    background: var(--bt-yellow);
+}
+
+.dropdown-cell {
+    display: flex;
+    flex-direction: row;
+    gap: 4px;
+    align-items: center;
+
+    select {
+        outline: none;
+        border: 0;
+        background-color: transparent;
+        color: var(--text-color);
+
+        option {
+            background: #000;
+            color: #fff;
+        }
+    }
+}
+
+.linear-picker.light-theme {
+    background: #fff;
+    color: #000;
+    border-color: #000;
+
+    .value-cell {
+        &.disabled {
+            color: #999;
+            background: #ccc;
+        }
+
+        &.selected {
+            background: var(--bt-yellow-background-light);
+        }
+    }
+
+    .state-button-cell.highlight:not(.disabled),
+    .button-cell.highlight:not(.disabled) {
+        background: var(--bt-yellow);
+    }
+
+    .dropdown-cell {
+        select {
+            color: #000;
+
+            option {
+                background: #fff;
+                color: #000;
+            }
+        }
+    }
+}
