+++ src/app/components/linear-picker/linear-picker-horizontal.component.html
@@ -1,0 +1,42 @@
+<div #picker class="linear-picker linear-picker-horizontal" [class.light-theme]="lightTheme()" [class.touch]="layoutService.isTouchInput()">
+    @if (title()) {
+    <div class="linear-picker-title">{{ title() }}</div>
+    }
+    <div class="linear-picker-values">
+        @for (choice of values(); let i = $index; track choice.value) {
+        @switch(choice.displayType) {
+        @case ('dropdown') {
+        <div class="value-cell dropdown-cell" [class.selected]="isSelected(choice)"
+            [class.highlight]="isHovered(choice) && !choice.disabled" [class.disabled]="choice.disabled"
+            [class.active]="choice.active">
+            <label [for]="'dropdown-horizontal-' + i">{{ choice.label }}</label>
+            <select [id]="'dropdown-horizontal-' + i" [title]="choice.label" (change)="handleDropdownChange($event, choice)">
+                @for (option of choice.choices; track option.value) {
+                <option [value]="option.value" [selected]="option.value === choice.value">{{ option.label }}</option>
+                }
+            </select>
+        </div>
+        }
+        @case ('label') {
+        <div class="value-cell label-cell">{{ choice.label }}</div>
+        }
+        @case ('state-button') {
+        <div class="value-cell state-button-cell" [class.selected]="isSelected(choice)"
+            [class.highlight]="isHovered(choice) && !choice.disabled" [class.disabled]="choice.disabled"
+            [class.active]="choice.active" (pointerenter)="setHoveredChoice(choice)" (pointerleave)="resetHovered()"
+            (pointerdown)="setHoveredChoice(choice)" (click)="handleChoiceClick($event, choice)">
+            {{ choice.label }}: {{ choice.shortLabel }}
+        </div>
+        }
+        @default {
+        <div class="value-cell button-cell" [class.selected]="isSelected(choice)"
+            [class.highlight]="isHovered(choice) && !choice.disabled" [class.disabled]="choice.disabled"
+            [class.active]="choice.active" (pointerenter)="setHoveredChoice(choice)" (pointerleave)="resetHovered()"
+            (pointerdown)="setHoveredChoice(choice)" (click)="handleChoiceClick($event, choice)">
+            {{ choice.shortLabel ?? choice.label }}
+        </div>
+        }
+        }
+        }
+    </div>
+</div>
