+++ src/app/components/linear-picker/linear-picker-horizontal.component.scss
@@ -1,0 +1,70 @@
+.linear-picker-horizontal {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    background: transparent;
+    border: 0;
+}
+
+.linear-picker-title {
+    box-sizing: border-box;
+    min-width: max-content;
+    padding: 3px 12px 4px;
+    border: 2px solid #fff;
+    border-bottom: 0;
+    background: #000;
+    color: #fff;
+    text-align: center;
+    text-transform: uppercase;
+    letter-spacing: 0.05em;
+    font-size: 0.8em;
+    line-height: 1.1;
+    font-weight: bold;
+}
+
+.linear-picker-values {
+    display: flex;
+    flex-direction: row;
+    width: max-content;
+    background: #000;
+    border: 2px solid #fff;
+    border-top: 0;
+}
+
+.value-cell {
+    padding: 8px 12px;
+    border-right: 1px solid rgba(255, 255, 255, 0.3);
+}
+
+.value-cell:last-child {
+    border-right: none;
+}
+
+.linear-picker-horizontal.touch .value-cell {
+    padding: 12px 16px;
+    font-size: 1em;
+}
+
+.value-cell.selected {
+    padding-left: 16px !important;
+    padding-right: 16px !important;
+}
+
+.linear-picker-horizontal.light-theme {
+    background: transparent;
+
+    .linear-picker-title {
+        background: #fff;
+        color: #000;
+        border-color: #000;
+    }
+
+    .linear-picker-values {
+        background: #fff;
+        border-color: #000;
+    }
+
+    .value-cell {
+        border-right-color: rgba(0, 0, 0, 0.3);
+    }
+}
