+++ src/app/components/linear-picker/linear-picker-horizontal.component.ts
@@ -1,0 +1,134 @@
+/*
+ * Copyright (C) 2025 The MegaMek Team. All Rights Reserved.
+ *
+ * This file is part of MekBay.
+ *
+ * MekBay is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License (GPL),
+ * version 3 or (at your option) any later version,
+ * as published by the Free Software Foundation.
+ *
+ * MekBay is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty
+ * of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU General Public License for more details.
+ *
+ * A copy of the GPL should have been included with this project;
+ * if not, see <https://www.gnu.org/licenses/>.
+ *
+ * NOTICE: The MegaMek organization is a non-profit group of volunteers
+ * creating free software for the BattleTech community.
+ *
+ * MechWarrior, BattleMech, `Mech and AeroTech are registered trademarks
+ * of The Topps Company, Inc. All Rights Reserved.
+ *
+ * Catalyst Game Labs and the Catalyst Game Labs logo are trademarks of
+ * InMediaRes Productions, LLC.
+ *
+ * MechWarrior Copyright Microsoft Corporation. MegaMek was created under
+ * Microsoft's "Game Content Usage Rules"
+ * <https://www.xbox.com/en-US/developers/rules> and it is not endorsed by or
+ * affiliated with Microsoft.
+ */
+
+import { ChangeDetectionStrategy, Component, input } from '@angular/core';
+import { LinearPickerBaseComponent } from './linear-picker-base.component';
+
+@Component({
+    selector: 'linear-picker-horizontal',
+    changeDetection: ChangeDetectionStrategy.OnPush,
+    imports: [],
+    templateUrl: './linear-picker-horizontal.component.html',
+    styleUrls: ['./linear-picker-common.scss', './linear-picker-horizontal.component.scss']
+})
+export class LinearPickerHorizontalComponent extends LinearPickerBaseComponent {
+    readonly align = input<'topleft' | 'left' | 'center' | 'top'>('center');
+
+    protected positionPicker(picker: HTMLDivElement): void {
+        const align = this.align();
+
+        if (align === 'topleft') {
+            this.positionPickerTopLeft(picker);
+        } else if (align === 'top') {
+            this.positionPickerTop(picker);
+        } else if (align === 'left') {
+            this.positionPickerLeft(picker);
+        } else {
+            this.recenterPicker(picker);
+        }
+    }
+
+    private recenterPicker(picker: HTMLDivElement): void {
+        const selectedCell = this.selectedCell(picker);
+        if (!selectedCell) {
+            this.centerPickerAtPosition(picker);
+            return;
+        }
+
+        this.centerPickerOnSelectedCell(picker, selectedCell);
+    }
+
+    private positionPickerTopLeft(picker: HTMLDivElement): void {
+        let leftPosition = Math.max(0, this.position().x);
+        picker.style.left = `${leftPosition}px`;
+        picker.style.top = `${this.position().y}px`;
+        picker.style.transform = 'translateY(-100%)';
+        picker.style.visibility = 'hidden';
+
+        requestAnimationFrame(() => {
+            const pickerRect = picker.getBoundingClientRect();
+            const viewportWidth = window.innerWidth;
+
+            if (pickerRect.right > viewportWidth) {
+                const overflow = pickerRect.right - viewportWidth;
+                leftPosition = Math.max(0, leftPosition - overflow);
+                picker.style.left = `${leftPosition}px`;
+            }
+            picker.style.visibility = 'visible';
+        });
+    }
+
+    private positionPickerTop(picker: HTMLDivElement): void {
+        picker.style.left = `${this.position().x}px`;
+        picker.style.top = `${this.position().y}px`;
+        picker.style.transform = 'translate(-50%, -100%)';
+        picker.style.visibility = 'hidden';
+
+        requestAnimationFrame(() => {
+            const pickerRect = picker.getBoundingClientRect();
+            const viewportWidth = window.innerWidth;
+            let adjustX = 0;
+
+            if (pickerRect.left < 0) {
+                adjustX = -pickerRect.left;
+            } else if (pickerRect.right > viewportWidth) {
+                adjustX = viewportWidth - pickerRect.right;
+            }
+
+            if (adjustX !== 0) {
+                picker.style.transform = `translate(calc(-50% + ${adjustX}px), -100%)`;
+            }
+            picker.style.visibility = 'visible';
+        });
+    }
+
+    private positionPickerLeft(picker: HTMLDivElement): void {
+        let leftPosition = Math.max(0, this.position().x);
+        picker.style.left = `${leftPosition}px`;
+        picker.style.top = `${this.position().y}px`;
+        picker.style.transform = 'translateY(-50%)';
+        picker.style.visibility = 'hidden';
+
+        requestAnimationFrame(() => {
+            const pickerRect = picker.getBoundingClientRect();
+            const viewportWidth = window.innerWidth;
+
+            if (pickerRect.right > viewportWidth) {
+                const overflow = pickerRect.right - viewportWidth;
+                leftPosition = Math.max(0, leftPosition - overflow);
+                picker.style.left = `${leftPosition}px`;
+            }
+            picker.style.visibility = 'visible';
+        });
+    }
+}
