+++ src/app/components/linear-picker/linear-picker-vertical.component.ts
@@ -1,0 +1,65 @@
+/*
+ * Copyright (C) 2025 The MegaMek Team. All Rights Reserved.
+ *
+ * This file is part of MekBay.
+ *
+ * MekBay is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License (GPL),
+ * version 3 or (at your option) any later version,
+ * as published by the Free Software Foundation.
+ *
+ * MekBay is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty
+ * of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU General Public License for more details.
+ *
+ * A copy of the GPL should have been included with this project;
+ * if not, see <https://www.gnu.org/licenses/>.
+ *
+ * NOTICE: The MegaMek organization is a non-profit group of volunteers
+ * creating free software for the BattleTech community.
+ *
+ * MechWarrior, BattleMech, `Mech and AeroTech are registered trademarks
+ * of The Topps Company, Inc. All Rights Reserved.
+ *
+ * Catalyst Game Labs and the Catalyst Game Labs logo are trademarks of
+ * InMediaRes Productions, LLC.
+ *
+ * MechWarrior Copyright Microsoft Corporation. MegaMek was created under
+ * Microsoft's "Game Content Usage Rules"
+ * <https://www.xbox.com/en-US/developers/rules> and it is not endorsed by or
+ * affiliated with Microsoft.
+ */
+
+import { ChangeDetectionStrategy, Component, signal } from '@angular/core';
+import { LinearPickerBaseComponent } from './linear-picker-base.component';
+
+@Component({
+    selector: 'linear-picker-vertical',
+    changeDetection: ChangeDetectionStrategy.OnPush,
+    imports: [],
+    templateUrl: './linear-picker-vertical.component.html',
+    styleUrls: ['./linear-picker-common.scss', './linear-picker-vertical.component.scss']
+})
+export class LinearPickerVerticalComponent extends LinearPickerBaseComponent {
+    readonly labelOnRight = signal<boolean>(false);
+
+    protected positionPicker(picker: HTMLDivElement): void {
+        const selectedCell = this.selectedCell(picker);
+        if (!selectedCell) {
+            this.centerPickerAtPosition(picker);
+        } else {
+            this.centerPickerOnSelectedCell(picker, selectedCell);
+        }
+        this.computeLabelSide(picker);
+    }
+
+    private computeLabelSide(picker: HTMLDivElement): void {
+        const rect = picker.getBoundingClientRect();
+        const labelWidth = 150;
+        const leftSpace = rect.left;
+        const rightSpace = window.innerWidth - rect.right;
+
+        this.labelOnRight.set(leftSpace < labelWidth && rightSpace > leftSpace);
+    }
+}
