--- src/app/components/linear-picker/linear-picker.component.scss
@@ -1,192 +1,0 @@
-.linear-picker {
-    font-family: 'Roboto', sans-serif;
-    position: fixed;
-    z-index: 9999;
-    -webkit-user-select: none;
-    user-select: none;
-    padding: 0;
-    background: #000;
-    color: #fff;
-    border: 2px solid #fff;
-    width: auto;
-    min-width: max-content;
-    white-space: nowrap;
-
-    &.horizontal {
-        display: flex;
-        flex-direction: row;
-    }
-
-    &.vertical {
-        display: flex;
-        flex-direction: column;
-    }
-}
-
-.value-cell.selected[data-title]::before {
-    position: absolute;
-    height: 100%;
-    min-height: 100%;
-    display: flex;
-    align-items: center;
-    padding-left: 12px;
-    padding-right: 12px;
-    border: 2px solid #000;
-    content: attr(data-title);
-    background-color: #000;
-    color: #fff;
-    white-space: nowrap;
-    text-transform: uppercase;
-    letter-spacing: 0.05em;
-    box-sizing: border-box;
-    font-weight: bold;
-}
-
-/* Vertical layout title positioning */
-.linear-picker.vertical .value-cell.selected[data-title]::before {
-    top: 0;
-    right: 100%;
-}
-
-.linear-picker.vertical.label-right .value-cell.selected[data-title]::before {
-    right: auto;
-    left: 100%;
-}
-
-/* Horizontal layout title positioning */
-.linear-picker.horizontal .value-cell.selected[data-title]::before {
-    bottom: 100%;
-    left: 0;
-    width: 100%;
-    min-width: 100%;
-    height: auto;
-    min-height: auto;
-    justify-content: center;
-}
-
-.value-cell {
-    position: relative;
-    font-size: 0.8em;
-    text-align: center;
-    cursor: pointer;
-    transition: background-color 0.15s ease-in-out;
-    flex-shrink: 0;
-}
-
-.value-cell.disabled {
-    cursor: not-allowed;
-    color: #666;
-    opacity: 0.6;
-}
-
-/* Vertical layout cell styling */
-.linear-picker.vertical .value-cell {
-    padding: 2px 8px;
-}
-
-.linear-picker.vertical.touch .value-cell {
-    padding: 3px 64px 3px 8px;
-    font-size: 0.9em;
-    text-align: right;
-}
-
-/* Horizontal layout cell styling */
-.linear-picker.horizontal .value-cell {
-    padding: 8px 12px;
-    border-right: 1px solid rgba(255, 255, 255, 0.3);
-}
-
-.linear-picker.horizontal .value-cell:last-child {
-    border-right: none;
-}
-
-.linear-picker.horizontal.touch .value-cell {
-    padding: 12px 16px;
-    font-size: 1em;
-}
-
-.value-cell.selected {
-    background: var(--bt-yellow-background-bright);
-    font-weight: bold;
-    z-index: 2;
-}
-
-/* Vertical layout selected cell */
-.linear-picker.vertical .value-cell.selected {
-    padding-top: 4px !important;
-    padding-bottom: 4px !important;
-}
-
-/* Horizontal layout selected cell */
-.linear-picker.horizontal .value-cell.selected {
-    padding-left: 16px !important;
-    padding-right: 16px !important;
-}
-
-.state-button-cell.highlight:not(.disabled),
-.button-cell.highlight:not(.disabled) {
-    background: var(--bt-yellow);
-}
-
-.dropdown-cell {
-    display: flex;
-    flex-direction: row;
-    gap: 4px;
-    align-items: center;
-
-    select {
-        outline: none;
-        border: 0;
-        background-color: transparent;
-        color: var(--text-color);
-
-        option {
-            background: #000;
-            color: #fff;
-        }
-    }
-}
-
-/* Light Theme styles */
-.linear-picker.light-theme {
-    background: #fff;
-    color: #000;
-    border-color: #000;
-
-    .value-cell {
-        &.selected[data-title]::before {
-            background: #fff;
-            color: #000;
-            border-color: #000;
-        }
-    
-        &.disabled {
-            color: #999;
-            background: #ccc;
-        }
-    
-        &.selected {
-            background: var(--bt-yellow-background-light);
-        }
-    }
-    .state-button-cell.highlight:not(.disabled),
-    .button-cell.highlight:not(.disabled) {
-        background: var(--bt-yellow);
-    }
-    
-    .dropdown-cell {
-    
-        select {
-            color: #000;
-    
-            option {
-                background: #fff;
-                color: #000;
-            }
-        }
-    }
-    
-    &.horizontal .value-cell {
-        border-right-color: rgba(0, 0, 0, 0.3);
-    }
-}
